+++ src/metric/metricscorer.hpp
@@ -14,7 +14,7 @@ class metricscorer {
 		void set_k(const unsigned int _k) { k = _k; }
 		unsigned int get_k() const { return k; }
 		virtual const char *whoami() const = 0;
-		virtual float compute_score(const qlist &ql) = 0;
+		virtual double compute_score(const qlist &ql) = 0;
 		virtual fsymmatrix *swap_change(const qlist &ql) = 0;
 		virtual void showme() { printf("\tscorer type = %s@%u\n", whoami(), k); };
 };
